move the old unmapped_plant_utils functions into that module
